requirements.txt
fastapi==0.111.0
googlemaps==4.10.0
h2==4.1.0
httptools==0.6.1
hume==0.7.0
importlib-metadata==7.2.0
inflect==7.3.0
ipykernel==6.29.4
itsdangerous==2.2.0
jsonpath-python==1.0.6
langchain-anthropic==0.1.15
langchain-community==0.2.5
langchain-mistralai==0.1.8
langchain-openai==0.1.9
pickleshare==0.7.5
pip-chill==1.0.3
pydantic-extra-types==2.8.2
pydantic-settings==2.3.3
retell-sdk==3.18.0
twilio==9.2.1
uvloop==0.19.0
watchfiles==0.22.0
